task_2/src/requirements.txt
pandas == 2.2.3
numpy == 2.1.3
matplotlib == 3.9.2
seaborn == 0.13.2
ipykernel == 6.29.5
pyyaml == 6.0.2
scikit-learn == 1.5.2
typing == 3.7.4.3
xgboost == 2.1.2
imbalanced-learn == 0.12.4